// src/classlogic/gunbreaker/enums.ts
export enum gunbreakerActionNames {
    KEEN_EDGE = 'Keen Edge',
    NO_MERCY = 'No Mercy',
    BRUTAL_SHELL = 'Brutal Shell',
    CAMOUFLAGE = 'Camouflage',
    DEMON_SLICE = 'Demon Slice',
    ROYAL_GUARD = 'Royal Guard',
    LIGHTNING_SHOT = 'Lightning Shot',
    SOLID_BARREL = 'Solid Barrel',
    BURST_STRIKE = 'Burst Strike',
    NEBULA = 'Nebula',
    DEMON_SLAUGHTER = 'Demon Slaughter',
    AURORA = 'Aurora',
    SUPERBOLIDE = 'Superbolide',
    SONIC_BREAK = 'Sonic Break',
    ROUGH_DIVIDE = 'Rough Divide',
    GNASHING_FANG = 'Gnashing Fang',
    SAVAGE_CLAW = 'Savage Claw',
    WICKED_TALON = 'Wicked Talon',
    BOW_SHOCK = 'Bow Shock',
    HEART_OF_LIGHT = 'Heart of Light',
    HEART_OF_STONE = 'Heart of Stone',
    CONTINUATION = 'Continuation',
    FATED_CIRCLE = 'Fated Circle',
    BLOODFEST = 'Bloodfest',
    BLASTING_ZONE = 'Blasting Zone',
    JUGULAR_RIP = 'Jugular Rip',
    ABDOMEN_TEAR = 'Abdomen Tear',
    EYE_GOUGE = 'Eye Gouge',
}

export enum gunbreakerStatusNames {
    NO_MERCY = 'No Mercy',
    READY_TO_RIP = 'Ready to Rip',
    READY_TO_TEAR = 'Ready to Tear',
    READY_TO_GOUGE = 'Ready to Gouge',
}

export enum gunbreakerDotNames {
    SONIC_BREAK = 'Sonic Break',
    BOW_SHOCK = 'Bow Shock',
}